+++ tests/test_cli.py
@@ -31,6 +31,8 @@ class ArgumentParser:
         self.splay = 0
         self.initial_run = False
         self.skip_render_on_undef = False
+        self.marker_set = '{{'
+        self.twopass_marker_set = None
 
 
 class TestCli(unittest.TestCase):
@@ -127,7 +129,6 @@ class TestCli(unittest.TestCase):
     def test_configure(self):
         args = ArgumentParser()
         args.config = 'config.json'
-        args.marker_set = '{{'
         open_mock = mock_open(read_data='{"searchlist": ["/foo"]}')
         with patch('e2j2.cli.open', open_mock):
             self.assertEqual(cli.configure(args)['searchlist'], ['/foo'])
@@ -192,7 +193,6 @@ class TestCli(unittest.TestCase):
 
     def test_run(self):
         args = ArgumentParser()
-        args.marker_set = '{{'
         # FIXME replace all args.filelist.split with lists see normal run
         # noop run
         args.noop = True
@@ -319,7 +319,6 @@ class TestCli(unittest.TestCase):
 
     def test_e2j2(self):
         args = ArgumentParser()
-        args.marker_set = '{{'
         args.stacktrace = True
 
         # Handle exceptions in config with stacktrace